Remove dead formatSchemaDate and fold formatEventDate into formatters
formatSchemaDate was an identity function with no production callers.
Delete it, move the remaining formatEventDate into formatters.ts (next to
the other string/data helpers), and drop the now-empty dateFormatter module.
